backend/pyproject.toml
[tool.poetry]
name = "backend-paw565pl"
version = "0.1.0"
description = ""
authors = ["Paw565pl <paw565pl@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^4.1"
djangorestframework = "^3.14.0"
djongo = "^1.3.6"
pymongo = {version = "3.13", extras = ["srv"]}
django-environ = "^0.11.2"
pytz = "^2023.3.post1"
django-filter = "^23.5"
drf-nested-routers = "^0.93.5"
tqdm = "^4.66.1"
djoser = "^2.2.2"
djangorestframework-simplejwt = { extras = ["crypto"], version = "^5.3.1" }
drf-spectacular = "^0.27.2"
django-cors-headers = "^4.3.1"

[tool.poetry.group.prod.dependencies]
whitenoise = { extras = ["brotli"], version = "^6.6.0" }
gunicorn = "^21.2.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
